requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scikit-learn-extra>=0.3.0
gower>=0.1.2
scipy>=1.10.0
openpyxl>=3.1.0
python-docx>=0.8.11
python-pptx>=0.6.21
